refactor: fix rewiev and logout in header
